Refactors GetPids
Moves GetPids inside cmd/delete.go as a helper function
Adds description
